+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changelog
 ------------------
 
 - using new prefixed image segmentation methods like `imgseg_from_bluechannel` instead of `to_bluechannel`
+- added `redis-predict-dp` filter for taking advantage of depth estimation docker images
 
 
 0.0.3 (2025-04-03)